fix: table component split
