+++ src/components/ConnectLine.tsx
@@ -6,13 +6,11 @@ import { ERDRelation, ERDTable } from '@/features/erd-project';
 import {
   getDrawLines,
   getDrawLinesMineMapping,
-  // getEndIDEFCircle,
   getEndIENotNullOneLine,
   getEndIENullableCircle,
   getEndIEOneLine,
   getManyLines,
   getStartEndPosition,
-  // getStartIDEFNullablePolygon,
   getStartIENotNullOneLine,
   getStartIENullableCircle,
   getStartIEOneLine,
@@ -20,25 +18,20 @@ import {
 } from '@/features/table-mapping';
 
 export function ConnectLine({
-  // crowFoot,
   tables,
   relation,
   tableDir,
   mineMapping,
 }: {
-  // crowFoot: boolean;
   tables: ERDTable[];
   relation: ERDRelation;
   tableDir: Map<ERDTable['id'], TableDirectionChild>;
   mineMapping: Map<ERDTable['id'], number>;
 }) {
-  return (
-    <>{SvgComponent(/* crowFoot, */ tables, relation, tableDir, mineMapping)}</>
-  );
+  return <>{SvgComponent(tables, relation, tableDir, mineMapping)}</>;
 }
 
 function SvgComponent(
-  // crowFoot: boolean,
   tables: ERDTable[],
   relation: ERDRelation,
   tableDir: Map<ERDTable['id'], TableDirectionChild>,
@@ -84,9 +77,7 @@ function SvgComponent(
 
     const drawLines = [];
     const drawCircles = [];
-    // const drawPolygons = [];
 
-    // if (crowFoot) {
     drawLines.push(getStartIEOneLine(fromDirection, updatedFrom));
 
     if (relation.type === 'one-to-many')
@@ -96,8 +87,8 @@ function SvgComponent(
 
     if (
       relation.multiplicity &&
-      relation.multiplicity.from &&
-      relation.multiplicity.from === 'optional'
+      relation.multiplicity.to &&
+      relation.multiplicity.to === 'optional'
     ) {
       drawCircles.push(getStartIENullableCircle(fromDirection, updatedFrom));
     } else {
@@ -106,26 +97,13 @@ function SvgComponent(
 
     if (
       relation.multiplicity &&
-      relation.multiplicity.to &&
-      relation.multiplicity.to === 'optional'
+      relation.multiplicity.from &&
+      relation.multiplicity.from === 'optional'
     ) {
       drawCircles.push(getEndIENullableCircle(toDirection, updatedTo));
     } else {
       drawLines.push(getEndIENotNullOneLine(toDirection, updatedTo));
     }
-    // } else {
-    //   drawCircles.push(getEndIDEFCircle(toDirection, updatedTo));
-
-    //   if (
-    //     relation.multiplicity &&
-    //     relation.multiplicity.from &&
-    //     relation.multiplicity.from === 'optional'
-    //   ) {
-    //     drawPolygons.push(
-    //       getStartIDEFNullablePolygon(fromDirection, updatedFrom),
-    //     );
-    //   }
-    // }
 
     navigateLines.forEach((line) => {
       lines.push(
@@ -165,22 +143,9 @@ function SvgComponent(
           cy={circle.y}
           stroke='#ededed'
           strokeWidth='1'
-          // fill={crowFoot ? '#2f2f2f' : '#ededed'}
         />,
       );
     });
-
-    // drawPolygons.forEach((polygon) => {
-    //   lines.push(
-    //     <polygon
-    //       key={Math.random().toString(36).slice(2)}
-    //       points={polygon.positions}
-    //       stroke='#ededed'
-    //       strokeWidth='1'
-    //       fill='#2f2f2f'
-    //     />,
-    //   );
-    // });
   }
 
   return lines;
