+++ src/components/implements-icon/MemoIcon.tsx
@@ -16,12 +16,12 @@ export function MemoIcon() {
         setEntity('memo');
       }}
     >
-      {Icon(fill)}
+      <Icon fill={fill} />
     </styles.container>
   );
 }
 
-function Icon(fill: string) {
+function Icon({ fill }: { fill: string }) {
   return (
     <svg
       xmlns='http://www.w3.org/2000/svg'
@@ -49,6 +49,7 @@ const styles = {
     justify-content: center;
     align-items: center;
   `,
+
   pathStyle: styled.path`
     &:hover {
       fill: #ddd;
