+++ src/components/implements-icon/PointerIcon.tsx
@@ -16,12 +16,12 @@ export function PointerIcon() {
         setCursor('pointer');
       }}
     >
-      {Icon(fill)}
+      <Icon fill={fill} />
     </styles.container>
   );
 }
 
-function Icon(fill: string) {
+function Icon({ fill }: { fill: string }) {
   return (
     <svg
       xmlns='http://www.w3.org/2000/svg'
@@ -53,6 +53,7 @@ const styles = {
     justify-content: center;
     align-items: center;
   `,
+
   pathStyle: styled.path`
     &:hover {
       fill: #ddd;
