+++ src/components/implements-icon/index.ts
@@ -3,3 +3,10 @@ export * from './ArrowIcon';
 export * from './MemoIcon';
 export * from './PointerIcon';
 export * from './TableIcon';
+export * from './IdentifyOneToOneIcon';
+export * from './IdentifyOneToManyIcon';
+export * from './IdentifyManyToManyIcon';
+export * from './NonIdentifyOneToOneIcon';
+export * from './NonIdentifyOneToManyIcon';
+export * from './NonIdentifyManyToManyIcon';
+export * from './MenuIcon';
