+++ src/components/index.ts
@@ -1,7 +1,6 @@
-export * from './TableMenu';
-export * from './RelationshipMenu';
+export * from './TableInformation';
+export * from './RelationshipInformation';
 export * from './ErdDrawTools';
-export * from './Table';
 export * from './ColumnEditMenu';
 export * from './Relationship';
 export * from './ConnectLine';
