+++ src/components/table/index.ts
@@ -1,0 +1,6 @@
+export * from './Table';
+export * from './TableMenu';
+export * from './Columns';
+export * from './FkColumns';
+export * from './PkColumns';
+export * from './PkFkColumns';
