+++ src/features/table-mapping/index.ts
@@ -1,1 +1,2 @@
 export * from './table-mapping.helper';
+export * from './table-mapping.type';
