+++ src/features/table-mapping/table-mapping.type.ts
@@ -1,0 +1,8 @@
+import { ERDTable } from '@/features/erd-project';
+
+export type TableDirectionChild = Map<
+  Direction,
+  { top: number; tableID: ERDTable['id'] }[]
+>;
+
+export type Direction = 'top' | 'bottom' | 'left' | 'right';
